refactor: 테스트 컨트롤러 이름 변경 및 confirmpayment 트랜잭션 적용
