+++ src/main/java/com/jelly/zzirit/domain/order/dto/response/PaymentConfirmResponse.java
@@ -1,5 +1,7 @@
 package com.jelly.zzirit.domain.order.dto.response;
 
+import com.jelly.zzirit.domain.order.entity.Order;
+
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @Schema(description = "결제 확인 응답 DTO")
@@ -13,4 +15,13 @@ public record PaymentConfirmResponse(
 
 	@Schema(description = "결제 금액", example = "12500")
 	int amount
-) {}
+
+) {
+	public static PaymentConfirmResponse from(Order order, String paymentKey) {
+		return new PaymentConfirmResponse(
+			order.getOrderNumber(),
+			paymentKey,
+			order.getTotalPrice().intValue()
+		);
+	}
+}
