+++ src/test/java/com/jelly/zzirit/domain/admin/controller/AdminControllerTest.java
@@ -27,7 +27,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @SpringBootTest
 @AutoConfigureMockMvc
-class AdminControllerRabbitTest extends RabbitTestMemberConfig {
+class AdminControllerTest extends RabbitTestMemberConfig {
 
     @Autowired
     private MockMvc mockMvc;
