+++ src/test/java/com/jelly/zzirit/domain/order/controller/pay/PaymentControllerTest.java
@@ -2,6 +2,6 @@ package com.jelly.zzirit.domain.order.controller.pay;
 
 import com.jelly.zzirit.global.support.AcceptanceRabbitTest;
 
-public class PaymentControllerRabbitTest extends AcceptanceRabbitTest {
+public class PaymentControllerTest extends AcceptanceRabbitTest {
 
 }
